Changed println() and throwing errors.

+++ src/main/kotlin/commands/RemoveFirstCommand.kt
@@ -7,7 +7,12 @@ class RemoveFirstCommand : Command (
     description = "Delete a first item in collection."
 ){
     override fun execute(args: List<String>, collectionManager: CollectionManager) {
+        if(collectionManager.isEmpty()) {
+            println("No element in the collection.")
+            return
+        }
         collectionManager.deleteByNumber(0)
+        println("First element removed.")
     }
 
 }
